+++ apps/server/package.json
@@ -30,6 +30,7 @@
     "@effect/language-service": "0.75.1",
     "@effect/vitest": "catalog:",
     "@t3tools/contracts": "workspace:*",
+    "@t3tools/shared": "workspace:*",
     "@t3tools/web": "workspace:*",
     "@types/bun": "^1.3.9",
     "@types/node": "^24.10.13",
