+++ packages/shared/package.json
@@ -1,0 +1,18 @@
+{
+  "name": "@t3tools/shared",
+  "version": "0.0.0-alpha.1",
+  "private": true,
+  "type": "module",
+  "exports": {
+    "./git": {
+      "types": "./src/git.ts",
+      "import": "./src/git.ts"
+    }
+  },
+  "scripts": {
+    "typecheck": "tsc --noEmit"
+  },
+  "devDependencies": {
+    "typescript": "^5.7.3"
+  }
+}
