+++ packages/shared/src/git.ts
@@ -1,0 +1,33 @@
+/**
+ * Sanitize an arbitrary string into a valid, lowercase git branch fragment.
+ * Strips quotes, collapses separators, limits to 64 chars.
+ */
+export function sanitizeBranchFragment(raw: string): string {
+  const normalized = raw
+    .trim()
+    .toLowerCase()
+    .replace(/['"`]/g, "")
+    .replace(/^[./\s_-]+|[./\s_-]+$/g, "");
+
+  const branchFragment = normalized
+    .replace(/[^a-z0-9/_-]+/g, "-")
+    .replace(/\/+/g, "/")
+    .replace(/-+/g, "-")
+    .replace(/^[./_-]+|[./_-]+$/g, "")
+    .slice(0, 64)
+    .replace(/[./_-]+$/g, "");
+
+  return branchFragment.length > 0 ? branchFragment : "update";
+}
+
+/**
+ * Sanitize a string into a `feature/…` branch name.
+ * Preserves an existing `feature/` prefix or slash-separated namespace.
+ */
+export function sanitizeFeatureBranchName(raw: string): string {
+  const sanitized = sanitizeBranchFragment(raw);
+  if (sanitized.includes("/")) {
+    return sanitized.startsWith("feature/") ? sanitized : `feature/${sanitized}`;
+  }
+  return `feature/${sanitized}`;
+}
